Updated README with initial information
This is what a commit message would look like.  Ideally, commit messages should provide some detail about what and why something in the source code changed.  This way you can read the commit message and get a quick understanding of the differences between the code as you may have remembered it and the code in its current state.