Updated template for slides for anyone wanting to use Reveal.js for slide decks that can be hosted via GitHub.
The file index.html contains a bunch of HTML that can be cut and pasted to construct the structure of anyone's slide deck.  There are also examples to show how to add presenter notes and things like that.  I can help with any of the HTML stuff related to this since I started using this library more and more for my own slide decks.