Adds an outline of Jason's section
Why:

* I'm way behind and you guys don't actually know what I'm talking
about!

This change addresses the need by:

* FIlling in the outline. I will probably abandon the presentation often
(or entirely) to work through an example of most of this with folks on
the side. I'm working on building an actual example in a private repo of
real code I'll end up using which I think will really illustrate this
part.